Don't fail to look up bare mangled names in apilookup.
The bug: if you look up a mangled name in apilookup without a DLL
name, and the name is not in the DB proper, it will not demangle it to
get its information.  This commit fixes this.

DecoratedDictionary had a stubbed-out get_api_definition(std::string
const &) const method.  Now it actually passes the name to the
demangler.