Revert the SWI Prolog 8.3.19 requirement; warn instead.
Allow building with SWI Prolog 8.2.0 or greater but warn at runtime when
a version less than 8.3.19 is being used.  At this time 8.3.19 is less
than two days old, and that's too heavy a requirement for building in
order to fix a relatively rare bug.

Also, include a few minor cosmetic and performance changes.