Pharos PIC hash update, 2025.01.30.
This branch is a work in progress.  It contains a major rewrite to
fn2hash that has major backward incompatibilities with the original
fn2hash.  PIC hashing, in particular, has been completely rewritten.
This is an unstable release, and future changes to this branch may also
change results without warning.

Fixes #267