Push poorly test NOP rewrite.
This code which Cory found while cleaning up an old directory, has a
complex history.  It was part of the effort to reconcile Pharos PIC
hashes with Ghidra PIC hashes, which uncovered that NOP detection in
Pharos was causing differences with Ghidra, and that it was actually
Pharos that was to blame.  Approximately the problem was that Pharos was
using ROSE's NOP detection which was faulty, and so Ghidra could not be
expected to replicate it.  As a result of Cory's NOP detection work
around the same time (see internal nops repo), this code was written to
possibly correct the issue.  But since we never decided to move forward
on the Ghidra portion, it became unclear whether the changes caused by
this code were worth the disruption and testing that would be required
before merging it into master (and also it was forgotten about).

My recommendation is to leave this commit on the top of the cfcpic
branch, and hopefully if the NOP detection raises it's head again in the
future, this code will serve as a useful guide on how to proceed.